Fix duplicated file
